refactor(hub-ui): split ModelLibrary panels into sibling file (≤800 lines); drop dead i18n key
- Extract ConnectedPanel, LocalPanel, NewProviderPanel, ModelChecklist + shared
  helpers/types into ModelLibraryPanels.tsx (634 lines)
- ModelLibrary.tsx now only owns the dialog shell + left provider rail (260 lines)
- Remove dead "lib.apiKeyPlaceholderConnected" key from en.ts + zh-TW.ts
